+++ snapservicefe/src/app/auth/register/components/RegisterForm.tsx
@@ -1,11 +1,18 @@
 'use client'
 
+import { getAPI } from '@/lib/axios'
 import { registerSupplier, registerUser } from '@/services/users/userService'
 import axios from 'axios'
 import { useRouter } from 'next/navigation'
 import { useRef, useState } from 'react'
 import { toast, ToastContainer } from 'react-toastify'
+const api = getAPI()
 
+type uploadResponse = {
+    url: string
+    key: string
+    imageUrl: string
+}
 export default function RegisterForm() {
     const [fullname, setFullname] = useState('')
     const [phone, setPhone] = useState('')
@@ -13,9 +20,64 @@ export default function RegisterForm() {
     const [password, setPassword] = useState('')
     const [role, setRole] = useState<'customer' | 'supplier'>('customer')
     const route = useRouter()
-    const [frontOfID, setFrontOfID] = useState('no file selected')
-    const [backOfID, setBackOfID] = useState('no file selected')
+    const [files, setFiles] = useState<File[]>([]);
+    const [message, setMessage] = useState("");
+    const [uploading, setUploading] = useState(false);
+    const ref = useRef<HTMLInputElement>(null);
 
+    const productId = 383;
+    const supplierId = 1; // tạm hardcode, mày có thể lấy từ props hoặc context
+
+    const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+        if (!e.target.files) return;
+        setFiles(Array.from(e.target.files));
+        setMessage("");
+    };
+
+    const handleUpload = async () => {
+        if (files.length === 0) {
+            setMessage("Chọn ít nhất 1 file đi bro!");
+            return;
+        }
+
+        setUploading(true);
+        setMessage("");
+
+        try {
+            // 1. Lấy mảng contentTypes để gửi lên backend lấy presigned URLs
+            const contentTypes = files.map((file) => file.type);
+
+            // 2. Gọi API backend lấy presigned URLs
+            const presignResponse = await api.post("/api/Upload/upload-images?isSupplierId=true",
+                {
+                    productId,
+                    productSlug: "", // nếu cần
+                    supplierId,
+                    categoryId: 0,
+                    contentTypes,
+                }
+            );
+
+            const uploads = presignResponse.data.data.uploads; // array { url, key, imageUrl }
+            // 3. Upload ảnh lên S3 song song
+            await Promise.all(
+                uploads.map((upload: any, idx: number) =>
+                    axios.put(upload.url, files[idx], {
+                        headers: { "Content-Type": files[idx].type },
+                    })
+                )
+            );
+
+            setMessage(`✅ Upload và confirm thành công ${files.length} ảnh!`);
+            setFiles([]);
+            return uploads; // Trả về mảng upload để sử dụng sau này
+        } catch (error: any) {
+            setMessage("Lỗi upload: " + (error.message || "Unknown error"));
+            console.error(error);
+        } finally {
+            setUploading(false);
+        }
+    };
 
     const handleSubmit = async (e: React.FormEvent) => {
         e.preventDefault()
@@ -29,13 +91,22 @@ export default function RegisterForm() {
                     }, 4000)
                 }
             } else {
-                const res = await registerSupplier({ fullname, phone, password, email })
+                const upload: uploadResponse[] = await handleUpload()
+                const res = await registerSupplier({
+                    fullname,
+                    phone,
+                    password,
+                    email,
+                    back_image: upload[0]?.imageUrl,
+                    front_image: upload[1]?.imageUrl,
+                });
                 if (res && res.username && res.email) {
                     toast.success("Đăng kí thành công")
                     setTimeout(() => {
                         route.push('/auth/login')
                     }, 4000)
                 }
+
             }
         } catch (error) {
             if (axios.isAxiosError(error)) {
@@ -104,25 +175,38 @@ export default function RegisterForm() {
                     />
                 </div>
                 {role === 'supplier' && (
-                    <div className='flex flex-col gap-4'>
-                        <div>
-                            <label className='block text-sm font-medium mb-1'>Front of ID card *</label>
-                            <div
-                                className="cursor-pointer bg-purple-600 text-white px-4 py-2 rounded-lg inline-block hover:bg-purple-700 transition"
-                            >
-                                Upload ID Card
-                            </div>
-                            <span className="ml-3 text-sm text-gray-600">{frontOfID}</span>
-                        </div>
+                    <div>
+                        <button
+                            type="button"
+                            onClick={() => ref.current?.click()}
+                            className="cursor-pointer bg-purple-600 text-white px-4 py-2 rounded-lg inline-block hover:bg-purple-700 transition"
+                        >
+                            Upload ID Card
+                        </button>
+                        <span className="ml-3 text-sm text-gray-600">
+                            {files.length > 0 ? `${files.length} file(s) selected` : "No files selected"}
+                        </span>
+                        <input
+                            type="file"
+                            multiple
+                            ref={ref}
+                            accept="image/*"
+                            onChange={handleFileChange}
+                            disabled={uploading}
+                            className="hidden"
+                        />
 
-                        <div>
-                            <label className='block text-sm font-medium mb-1'>Back of ID card *</label>
-                            <div
-                                className="cursor-pointer bg-purple-600 text-white px-4 py-2 rounded-lg inline-block hover:bg-purple-700 transition"
-                            >
-                                Upload ID Card
-                            </div>
-                            <span className="ml-3 text-sm text-gray-600">{backOfID}</span>
+                        {message && <p>{message}</p>}
+
+                        <div style={{ display: "flex", gap: 8, marginTop: 12 }}>
+                            {files.map((file, idx) => (
+                                <img
+                                    key={idx}
+                                    src={URL.createObjectURL(file)}
+                                    alt={`preview-${idx}`}
+                                    style={{ width: 100, height: 100, objectFit: "cover", borderRadius: 6 }}
+                                />
+                            ))}
                         </div>
                     </div>
                 )}
