+++ snapservicefe/src/model/request/userRequest.ts
@@ -5,6 +5,15 @@ export interface registerRequest {
     password: string;
 }
 
+export interface registerSupplierRequest {
+    fullname: string;
+    email: string;
+    phone: string;
+    password: string;
+    back_image: string;
+    front_image: string;
+}
+
 export interface loginRequest {
     emailOrPhone: string;
     password: string;
